fix getting x

+++ ft_va_arg_wrapper.c
@@ -31,7 +31,7 @@ char	*ft_get_arg(char type, va_list *args, int *has_null)
 	else if (type == 'u')
 		param_str = ft_long_itoa(va_arg(*args, unsigned int));
 	else if (type == 'x' || type == 'X')
-		param_str = ft_convert_to_hex(va_arg(*args, long), type == 'X');
+		param_str = ft_convert_to_hex(va_arg(*args, unsigned int), type == 'X');
 	else if (type == 'p')
 		param_str = ft_va_arg_p(args);
 	else if (type == '%')
